refactor: change component files to use barrel exports
