--- src/components/GameBoard.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const GameBoard = () => {
-  return (
-    <div className='gameboard'>
-      Game Board
-    </div>
-  )
-}
-
-export default GameBoard
