--- src/components/Title.tsx
@@ -1,5 +1,0 @@
-const Title = () => {
-  return <h1 className="text-blue-500">Memory Game</h1>;
-};
-
-export default Title;
