+++ src/components/game-board.tsx
@@ -1,0 +1,5 @@
+const GameBoard = () => {
+  return <div className='gameboard'>Game Board</div>;
+};
+
+export { GameBoard };
