+++ src/components/index.ts
@@ -1,1 +1,4 @@
-export * from '.';
+export * from './game-board';
+export * from './score-board';
+export * from './setup';
+export * from './title';
