+++ src/components/title.tsx
@@ -1,0 +1,5 @@
+const Title = () => {
+  return <h1 className='text-blue-500'>Memory Game</h1>;
+};
+
+export { Title };
